lakemind-api-service/requirements.txt
# LakeMind API Service Dependencies
# Resolved on Python 3.14 — all pins verified compatible
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
beautifulsoup4==4.14.3
cachetools==7.0.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
databricks-connect==15.4.21
databricks-sdk==0.105.0
databricks-sql-connector==4.2.5
debugpy==1.8.20
et-xmlfile==2.0.0
fastapi==0.136.0
google-auth==2.49.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
idna==3.12
lz4==4.4.5
Mako==1.3.11
MarkupSafe==3.0.3
marshmallow==4.3.0
networkx==3.6.1
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.1
pandas==2.3.3
pendulum==3.2.0
protobuf==6.33.6
psycopg[binary]==3.3.3
py4j==0.10.9.7
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1-modules==0.4.2
pybreaker==1.4.1
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.13.3
pydantic-core==2.46.3
PyJWT==2.12.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-http-client==3.3.7
python-multipart==0.0.26
pytz==2026.1.post1
requests==2.33.1
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
thrift==0.20.0
time-machine==3.2.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.45.0
